Optimization update
View frustum culling on Quad tree added. I have also added a top down view for the first render pass so we can in real time see what quads are being rendered. In the second render call we just render the scene as usual. Loading big maps take a lot time, especially when you run in Debug mode, so I advice you to run it in release. There is also a weird bug with the camera when the map is too big and you travel too far it suddenly gets inverted. I have no idea why. But on a map of the size of 2^7 + 1, ther is no problems and as usual play around with the parameters for map creation. Important note: Do not make mountains too big, then the view frustum culling will cull things that should not be culled because of the height value (it will cull depending of the cube radius). Meaning that if you have a small map, but high peaks and you travel upward, if the camera position is above the radius (radius is depending the map/cube size) the entire quad will be culled since its now "outside" the frustum.